requirements.txt
streamlit
langchain
openai
langchain-community
python-dotenv
pymupdf
